requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiofiles==0.7.0
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
cryptography==35.0.0; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.70.0
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gcloud==0.18.3
googleapis-common-protos==1.53.0; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
httplib2==0.20.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
jinja2==3.0.2
jwcrypto==1.0
line-bot-sdk==1.20.0
markupsafe==2.0.1; python_version >= '3.6'
oauth2client==4.1.3
protobuf==3.18.1; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.8.2; python_full_version >= '3.6.1'
pymongo==3.12.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrebase4==4.5.0
python-jwt==3.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.9.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.16.0; python_version >= '3.6'
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.15.0
wrapt==1.13.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiofiles==0.7.0
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
cryptography==35.0.0; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.70.0
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gcloud==0.18.3
googleapis-common-protos==1.53.0; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
httplib2==0.20.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
jinja2==3.0.2
jwcrypto==1.0
line-bot-sdk==1.20.0
markupsafe==2.0.1; python_version >= '3.6'
oauth2client==4.1.3
protobuf==3.18.1; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.8.2; python_full_version >= '3.6.1'
pymongo==3.12.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrebase4==4.5.0
python-jwt==3.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.9.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.16.0; python_version >= '3.6'
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.15.0
wrapt==1.13.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
